When adding a statement, also pick up on nested labels.
Adding a statement to a CFG checks if the statement is a label, and adds
it to a CFG-global list of labels if this is true. However, this misses
labels when they are nested within other statements (e.g., if state-
ments). With this change, also nested labels are added to the CFG-global
list of labels.