Refactor `predicate` to include check for other attrs
This change includes parsing the function as an `untyped::AnyFnItem`
instead of `syn::Item`, some more checks beside the main one for other
prusti attributes and as a side effect improving the error messages to
only list the actual errors we care about, not syntactic problems when
parsing e.g. a static item as function.