Work around AssertKind::description panicing for BoundsCheck
AssertKind::BoundsCheck panics on calling AssertKind::display, and the
docs say that the caller is supposed to handle those earlier[0]. So we
use the Debug impl instead, which can handle BoundsCheck.

At some point we want to switch to using `AssertKind::fmt_assert_args`,
but currently that is not public in rustc.

[0] https://doc.rust-lang.org/stable/nightly-rustc/rustc_middle/mir/enum.AssertKind.html#impl