Extern Spec Improvements (#1188)
* allow default bodies for trait predicates

* support more types in #[refine_trait_spec]

* correctly handle match guards

They're encoded as false edges, but with only one state, having both arms point to the same basic block. This violates a previously-made assumption, but the code never relied on that anyway.

* make body of extern spec methods type check return type

* move extract_type_path to common

* allow extern specs on types that aren't simple type paths

* remove unnecessary refs/derefs

thanks, clippy!

* update tests

* update expected output for desugaring test

* update error-expecting test for abstract predicates

* rewrite self type to type rather than type path

no more extract_type_path; i was wrong about how it's used

* add TODO about trait predicates behavioral subtyping

* formatting

wish this crate was part of the auto-formatted ones

* fix newly-identified type mismatch in old test

* fix newly-identified bounds mismatch in prusti-std

* share & improve extern spec function codegen

- now shared between methods and free functions, after the previous duplication threw me off thinking that my changes were not being applied…
- the panic is back, though in a different form to preserve type checking of the result
- generic args are now explicitly forwarded to the original method, for things like `core::mem::size_of::<T>()`

* fix path being mutated across sibling modules

this fails for e.g. `mod core { mod cmp { ... } mod mem { ... } }`, trying to use `core::cmp::mem` for the items in `core::mem`

* update tests

* formatting

* move lifetime explicitization logic

to the shared function, where it belongs

* remove unnecessary unreachable!() sanity check

the code this generates is never executed under any circumstances, so this check is completely unnecessary

* add unit tests for generics forwarding