Change std::collections::{HashMap, HashSet} to fxhash::{FxHashMap, FxHashSet}
FxHash was already in use in parts of the prusti-common crate (e.g. in program_normalization.rs),
now every file in the crate uses the FxHash versions of HashMap and HashSet.

Many functions in the prusti-common crate create a new Map or Set each time
they are called, and the standard HashMap and HashSet have to create a secure seed
every time. The FxHash variants should improve performance of these functions.

A clippy lint was also added to warn about future uses of the standard
collections HashMap and HashSet