Refactor code and change to sort_unstable
Changed code from for loops to map + collect to reduce number of vector reallocations.
Changed Vec::sort() to Vec::sort_unstable(), because it is faster and stability is not
needed. Most of the sorted vectors are created from a hashmap, so there are no duplicate
entries whose order could be changed by the sorting.