Completion of issue 711 and general code improvements (#1314)
* Changing sets and maps to use the same hasher

* Fixed typo

* Change to FxHash

* Refactor code and change to sort_unstable

Changed code from for loops to map + collect to reduce number of vector reallocations.
Changed Vec::sort() to Vec::sort_unstable(), because it is faster and stability is not
needed. Most of the sorted vectors are created from a hashmap, so there are no duplicate
entries whose order could be changed by the sorting.

* Preallocate vector capacity

* Preallocate vector capacity

* Remove unnecessary work

The old code created the viper encodings for all the methods, even
if they get deleted again afterwards if config::verify_only_preamble()
is true. Now it will only do the work if the option is not set to true

* Fixed formating

* Inline variables in format string

* Switch to unstable sort

* Switch to FxHash

* Switch to rustc-hash

* Fix formatting

* Remove unused file

* Reduce error message duplication

* Switch to rustc-hash

* Improve error reporting

* Add clippy lint for disallowed types

* Fix incorrect format string

Co-authored-by: Aurel <Aurel300@users.noreply.github.com>

* Disable rustup self update

---------

Co-authored-by: Aurel <Aurel300@users.noreply.github.com>