Implement export/import of quantifiers
When importing modules which contain specifications, these specs are
also imported. The spec functions are serialized and written to a .specs
file when the module is verified, and can be loaded from this file where
the module is imported.
However, if there are quantifiers in the exported specs, this did not
work before. Quantifiers are represented as closures, and are not
included the spec functions themselves, they need to be exported
separately. This commit implements that.