Expire cache when Prusti version changes
The previous commit changes the format of the .specs file.
If you still have old .specs files in your target folder and run the new
Prusti, it will panic trying to parse the old .specs file.
This can be solved by deleting the target folder, but that is not a nice
user experience. We would like everything to be re-verified when Prusti
is updated.

Cargo uses various pieces of information to determine whether a
recompile is needed, among them the rustc version. Cargo runs
`rustc -vV` to obtain that. By replacing rustc with prusti-rustc, and
outputting the Prusti version when running `prusti-rustc -vV`,
we can achieve our goal. Everything is re-verified when the Prusti
version changes.

Prusti previously set RUSTC_WRAPPER instead of RUSTC, but cargo does not
use the wrapper when querying the version. I don't know why
RUSTC_WRAPPER was used instead of RUSTC. Switching to RUSTC does not
seem to cause problems.
However, in prusti-rustc, we can no longer look for a rustc argument to
see if we are under cargo, so instead I set the new PRUSTI_CARGO env to
signal this explicitly.